@@ -1,13 +1,19 @@
 import store from "./redux/redux-store";
 import "./index.css";
 import App from "./App";
-import ReactDOM from "react-dom";
-
+import { createRoot } from "react-dom/client";
 import { Provider } from "react-redux";
 
-ReactDOM.render(
+const root = document.getElementById("root");
+
+if (!root) {
+  throw new Error("root not found");
+}
+
+const container = createRoot(root);
+
+container.render(
   <Provider store={store}>
     <App />
-  </Provider>,
-  document.getElementById("root")
+  </Provider>
 );
